Add files via upload
Here are the files I am using for the project.  Note: The code for the "Movie Statistics by Year" page is as complete as far as I can get it until the data is finalized, as this will dictate the code that is written.  Also, there will aesthetic changes needed down the road.